update  solve PR-CI-Kunlun-KP-Build
